# [PoC] Introduce new flag `SpecCacheDisabled` & Parse only the requires BTF types