--- a/btf/kernel.go
+++ b/btf/kernel.go
@@ -29,6 +29,21 @@ func FlushKernelSpec() {
 	kernelBTF.modules = make(map[string]*Spec)
 }
 
+// LoadKernelSpecWithOptions returns the current kernel's BTF information.
+//
+// Defaults to /sys/kernel/btf/vmlinux and falls back to scanning the file system
+// for vmlinux ELFs. Returns an error wrapping ErrNotSupported if BTF is not enabled.
+//
+// The provided options are used to configure the BTF loading process.
+func LoadKernelSpecWithOptions(opts *SpecOptions) (*Spec, error) {
+	if kernelBTF.kernel != nil {
+		return kernelBTF.kernel.Copy(), nil
+	}
+
+	spec, _, err := loadKernelSpec(opts)
+	return spec, err
+}
+
 // LoadKernelSpec returns the current kernel's BTF information.
 //
 // Defaults to /sys/kernel/btf/vmlinux and falls back to scanning the file system
@@ -49,7 +64,7 @@ func LoadKernelSpec() (*Spec, error) {
 		return spec.Copy(), nil
 	}
 
-	spec, _, err := loadKernelSpec()
+	spec, _, err := loadKernelSpec(nil)
 	if err != nil {
 		return nil, err
 	}
@@ -93,7 +108,7 @@ func LoadKernelModuleSpec(module string) (*Spec, error) {
 	return spec.Copy(), nil
 }
 
-func loadKernelSpec() (_ *Spec, fallback bool, _ error) {
+func loadKernelSpec(opts *SpecOptions) (_ *Spec, fallback bool, _ error) {
 	if platform.IsWindows {
 		return nil, false, internal.ErrNotSupportedOnOS
 	}
@@ -102,7 +117,7 @@ func loadKernelSpec() (_ *Spec, fallback bool, _ error) {
 	if err == nil {
 		defer fh.Close()
 
-		spec, err := loadRawSpec(fh, internal.NativeEndian, nil)
+		spec, err := loadRawSpec(fh, internal.NativeEndian, nil, opts)
 		return spec, false, err
 	}
 
@@ -112,7 +127,7 @@ func loadKernelSpec() (_ *Spec, fallback bool, _ error) {
 	}
 	defer file.Close()
 
-	spec, err := LoadSpecFromReader(file)
+	spec, err := LoadSpecFromReaderWithOptions(file, opts)
 	return spec, true, err
 }
 
@@ -132,7 +147,7 @@ func loadKernelModuleSpec(module string, base *Spec) (*Spec, error) {
 	}
 	defer fh.Close()
 
-	return loadRawSpec(fh, internal.NativeEndian, base)
+	return loadRawSpec(fh, internal.NativeEndian, base, nil)
 }
 
 // findVMLinux scans multiple well-known paths for vmlinux kernel images.