… btf/btf_test.go
… btf/fuzz_test.go
… btf/kernel_test.go
… btf/marshal_test.go
… btf/types_test.go