Based on https://github.com/cilium/ebpf/pull/1589

This is a proof-of-concept. The code is not ready for merging but it shows it is possible to significantly reduce the memory consumption (by 27MB).

---

This PR aims to lower the memory footprint when using the `cilium/ebpf` library. This is achieved in two ways:
1. Reducing memory usage: Adding a new flag which disables caching of the BTF spec
2. Parsing only the needed BTF types

The tradeoff is for lowering the memory footprint is of course performance while loading eBPF programs etc...

In this PoC when using the new `SpecCacheDisabled` flag (1) it will also automatically check which BTF types are needed and load/parse only those (2)

---

Benchmarks:
1. `ParseVmlinux` is always the base for both tests
2. The `new.txt` Inspektor gadget run is with the types from https://github.com/cilium/ebpf/pull/1755/#issuecomment-2820919961
3. The `new.txt` `ParseVmlinux` run is called `BenchmarkParseVmlinuxWithoutFilter` in this PR
```
goos: linux
goarch: amd64
pkg: github.com/cilium/ebpf/btf
cpu: 11th Gen Intel(R) Core(TM) i7-11370H @ 3.30GHz
                     │   old.txt   │               new.txt                │
                     │   sec/op    │   sec/op     vs base                 │
ParseVmlinux-8         29.28m ± 3%   60.62m ± 8%  +107.06% (p=0.000 n=10)
InspektorGadget-8      29.28m ± 3%   12.94m ± 3%   -55.79% (p=0.000 n=10)
geomean                29.28m        28.01m         -4.32%

                     │   old.txt    │                new.txt                │
                     │     B/op     │     B/op      vs base                 │
ParseVmlinux-8         24.43Mi ± 0%   53.58Mi ± 0%  +119.37% (p=0.000 n=10)
InspektorGadget-8      24.43Mi ± 0%   10.89Mi ± 0%   -55.40% (p=0.000 n=10)
geomean                24.43Mi        24.16Mi         -1.08%

                     │   old.txt   │               new.txt               │
                     │  allocs/op  │  allocs/op   vs base                │
ParseVmlinux-8         271.9k ± 0%   365.1k ± 0%  +34.30% (p=0.000 n=10)
InspektorGadget-8      271.9k ± 0%   110.1k ± 0%  -59.51% (p=0.000 n=10)
geomean                271.9k        200.5k       -26.26%
```