@burak-ok 2025-04-17
> Seems like vmlinux spec keeps being a problem! Just checking: when you say memory usage you mean heap at idle? Calling https://pkg.go.dev/github.com/cilium/ebpf/btf#FlushKernelSpec does not help?

Yes, that would help for having a lower heap usage after starting the program. But if one sets a low memory limit in a pod spec, one also needs to avoid high memory while initializing -> while loading every program.

Furthermore with `FlushKernelSpec` we need to find and identify every possible call into the `ebpf` library which might parse and save the while KernelSpec. With a flag we can set the flag and forget about flushing the kernel spec.

> How would you determine which types to parse from vmlinux?

For that we are reading the all `relo.TypeNames` out of the program where the relocations are getting applied: https://github.com/cilium/ebpf/pull/1755/files#diff-981ef293a9c93614e843135eb5b207f951babe5604aee33c72f66567ccaa01de