@lmb 2025-04-23
@burak-ok can you take a look at https://github.com/cilium/ebpf/pull/1763? Doesn't address FlushKernelSpec, but I think that might be better done by relying on `weak` in Go 1.24.