@burak-ok 2025-04-25
> @burak-ok can you take a look at #1763?

I added roughly the same benchmarks that you posted for this PR in the top post.
Comparing these results your PR saves roughly the same amount of memory and allocations for the `InspektorGadget` benchmark.
But the `ParseVmlinux` Benchmark shows that my Draft PR has more disadvantages for the general usecase of reading and parsing every type.

I'll take a deeper look into your PR some time later, thanks for opening it.

> Doesn't address FlushKernelSpec, but I think that might be better done by relying on `weak` in Go 1.24.

With having memory limits I think this would be the best case scenario for us.