@lmb 2025-05-07
The new lazy BTF code is in. @burak-ok could you try the code and report back how much of a difference it makes?